ai powered pdf2/AI-powered-pdf-reader/requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
numpy
Pillow
